requirements.txt
# HonyGo - 极致轻量化自动鼠标点击工具
# Python依赖包列表
# 优化版本 - 基于实际使用情况分析

# GUI框架
PySide6>=6.6.0

# 图像处理
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# OCR相关
easyocr>=1.7.0
torch>=2.0.0
tensorflow>=2.13.0

# GPU监控和加速
pynvml>=11.5.0
GPUtil>=1.4.0

# 网络和API
requests>=2.31.0
flask>=2.3.0
flask-cors>=4.0.0
urllib3>=2.0.0

# 系统交互
psutil>=5.9.0
pyautogui>=0.9.54
pywin32>=306

# 任务调度
schedule>=1.2.0

# 配置和验证
jsonschema>=4.17.0

# 包管理和元数据
importlib-metadata>=6.0.0

# 缺失的依赖 - 基于代码分析添加
keyboard>=0.13.0
watchdog>=3.0.0

# 注意：coloredlogs和pynput已移除（未在代码中使用）

# 注意：以下依赖已从原requirements.txt中移除（未在代码中使用）：
# - nvidia-ml-py (未使用)
# - pytest (开发依赖，可选)
# - pytest-qt (开发依赖，可选)
# - pyyaml (未使用)
# - torchvision (未使用)

# tkinter 是Python标准库的一部分，无需在requirements.txt中指定